Use Buffer.from instead of toBuffer() in save(modelArtifacts)
Replace call to `io_utils.toBuffer(arrayBuffer)` with
`Buffer.from(arrayBuffer)`.

The difference is that `Buffer.from(arrayBuffer)` does not perform
copying of the underlying data and just sets up a view on top on the
underlying ArrayBuffer.

It's safe to do that when the underlying data is not changed, and it is
faster.