toArrayBuffer: minor code cleanup and optimization
This does not change observed toArrayBuffer() behavior.

Change toArrayBuffer(Buffer) to directly slice the underlying
ArrayBuffer object to return a copy of the data.

Buffer is a subclass of Uint8Array on all supported Node.js versions,
so we can use `.buffer`, `.byteLength` and `.byteOffset` properties.

Internally, this removes an extra zero-fill from the ArrayBuffer.

toArrayBuffer(Buffer[]) is only optimized to use native Buffer#copy()
method instead of manual byte-by-byte copying.

Also replace .forEach() with for-of (mostly a style change).

Refs: https://nodejs.org/api/buffer.html#buffer_buf_copy_target_targetstart_sourcestart_sourceend