toBuffer: use Buffer.from(Uint8Array) to reduce filling
This does not change observed toBuffer() behavior.

Buffer.from(Uint8Array) is supported in Node.js since a long time,
and is explicitly checked for since Node.js v8.0.0.

Internally, this removes one extra fill from the buffer: this way,
it's not zero-filled on creation, but instead is directly filled with
values from the temporary Uint8Array.

Note that a temporary Uint8Array is still created (without copying),
because direct Buffer.from(ArrayBuffer) would have returned a zero-copy
view on top of the given ArrayBuffer object.

 * `Buffer.from(ab: ArrayBuffer)` does not copy the underlying data and
   could be used safely only while `ab` is guranteed to be not modified.

 * `toBuffer(ab: ArrayBuffer)` copies data once to a new Buffer instance.

Refs: https://github.com/nodejs/node/commit/beca3244e2792bc4257f480b138c1b603de3781c
Refs: https://nodejs.org/api/buffer.html#buffer_class_method_buffer_from_buffer