file_system: avoid relying on in-advance fs.exists check in dir creation
This avoids possible race conditions, e.g. when more than one
parallel async .save() operation is called (from a single or several
processes) sharing a common prefix which does not yet exist.

Instead, createOrVerifyDirectory() now just tries creating the dir,
and if that fails -- checks _why_ that failed, if it is anything
other than the directory already existing -- throws an error, if
it's the dir already existing -- skips that and continues.

Refs: https://nodejs.org/api/fs.html#fs_fs_exists_path_callback